Initial Cleanup
Moved some functions around to increase clarity of action, rather than
having the whole program in one file. Next step is taking the handler
into a separate file, since it's a class anyway. That'll additionally
allow me to get rid of the final global variables, the textfields.
